% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation.R
\name{print.conditional_indirect_effect}
\alias{print.conditional_indirect_effect}
\title{Print a moderated mediation decomposition}
\usage{
\method{print}{conditional_indirect_effect}(x, ...)
}
\arguments{
\item{x}{A \code{conditional_indirect_effect} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a moderated mediation decomposition
}
