% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation.R
\name{print.conditional_slopes}
\alias{print.conditional_slopes}
\title{Print conditional (simple) slopes}
\usage{
\method{print}{conditional_slopes}(x, ...)
}
\arguments{
\item{x}{A \code{conditional_slopes} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print conditional (simple) slopes
}
