% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{print.cssem_association}
\alias{print.cssem_association}
\title{Print an associational CS-SEM structural fit}
\usage{
\method{print}{cssem_association}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_association} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an associational CS-SEM structural fit
}
