% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.cssem_fit}
\alias{print.cssem_fit}
\title{Print a CS-SEM fit}
\usage{
\method{print}{cssem_fit}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_fit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CS-SEM fit
}
