% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{print.cssem_mediation}
\alias{print.cssem_mediation}
\title{Print an associational CS-SEM mediation decomposition}
\usage{
\method{print}{cssem_mediation}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_mediation} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an associational CS-SEM mediation decomposition
}
