% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.cssem_model}
\alias{print.cssem_model}
\title{Print a CS-SEM measurement model}
\usage{
\method{print}{cssem_model}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_model} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CS-SEM measurement model
}
