% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation.R
\name{print.cssem_moderated_mediation}
\alias{print.cssem_moderated_mediation}
\title{Print a moderated mediation decomposition}
\usage{
\method{print}{cssem_moderated_mediation}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_moderated_mediation} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a moderated mediation decomposition
}
