% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{print.cssem_routing}
\alias{print.cssem_routing}
\title{Print a Path Routing Table}
\usage{
\method{print}{cssem_routing}(x, ...)
}
\arguments{
\item{x}{A \code{cssem_routing} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Path Routing Table
}
