% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.fit_states}
\alias{print.fit_states}
\title{Print a CS-SEM fit}
\usage{
\method{print}{fit_states}(x, ...)
}
\arguments{
\item{x}{A \code{fit_states} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a CS-SEM fit
}
