% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{print.indirect_effect}
\alias{print.indirect_effect}
\title{Print an associational CS-SEM mediation decomposition}
\usage{
\method{print}{indirect_effect}(x, ...)
}
\arguments{
\item{x}{A \code{indirect_effect} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an associational CS-SEM mediation decomposition
}
